Utilにテスト導入してみようかと思い立った

+++ app/src/main/java/com/ict/mito/gootravel/spot/model/livrdata/OrientationLiveData.kt
@@ -95,7 +95,7 @@ class OrientationLiveData(
             remapCoordinateArray,
             orientationAngles
         )
-        val angle = rad2deg(orientationAngles[0])
-        orientationAngles[0] = normalizeRange(angle)
+        val angle = rad2deg(orientationAngles[0].toDouble())
+        orientationAngles[0] = normalizeRange(angle).toFloat()
     }
 }
